Remove clickFactor from overallScore calculation
The clickFactor variable was removed from the calculation of overallScore in the script.js. This change results in a new formula that no longer includes the number of clicks in the calculation of the score. The aim is to focus the user's score on time management, difficulty level, and points per click only.

Took 8 minutes