Change `touchstart` to `touchend` event and add audio error handling
Switched touch event from `touchstart` to `touchend` to respond after touch action is complete. Additionally, error handling was added to the `startAudio` method to catch potential audio play failures and a conditional check was put in place to initiate audio if sound is off when restarting the game.

Took 1 hour 1 minute


Took 46 seconds