common: Refactor time elapsed and score calculation in script.js
The previous method of calculating time elapsed and the score multiplier was corrected. The variable timeElapsed is now a constant and does not multiply by 10, removing an unnecessary step. The timeFactor will now be set to zero if it becomes negative, enforcing a positive value for the overall score calculation.

Took 7 minutes


Took 18 seconds